app/requirements.txt
Flask
mysql-connector-python
pandas
scikit-learn
matplotlib
streamlit
python-dotenv
